Add RSC support via dual entry points and apply review fixes
Split library into server-safe (react-slots-server.tsx) and client
(react-slots.tsx with 'use client') entry points. Server module
exports getSlots and types with no hooks. Client re-exports
everything and adds withSlots, SlottedComponent, useSlots.

Review fixes applied:
- Fix __DEV__ returning true when process.env.NODE_ENV is absent
- Make __DEV__ module-private (was accidentally exported)
- Use React.Fragment instead of Symbol.for('react.fragment')
- Remove dead constructorMap entry for namespaced slots
- Remove unreachable multi-slot branching (array always pre-initialized)

Updates build.js for dual bundles, package.json exports for ./server
subpath, and README with server component docs and all v0.3.0 changes.
